fingerprint handling changes, fix test report on error, simplify superagent usage
