+++ integration-test/.prettierrc
@@ -6,5 +6,13 @@
     "bracketSameLine": false,
     "semi": true,
     "trailingComma": "none",
-    "printWidth": 110
+    "printWidth": 110,
+    "importOrder": [
+        "<THIRD_PARTY_MODULES>",
+        "^lib/(.*)$",
+        "^[../]",
+        "^[./]"
+    ],
+    "importOrderSeparation": true,
+    "importOrderSortSpecifiers": true
 }
