+++ integration-test/package.json
@@ -9,10 +9,12 @@
     "mkcert:dev": "mkcert create-cert --ca-cert ../service/certs/ca.crt --ca-key ../service/certs/ca.key --cert ../service/certs/scytta.crt --key ../service/certs/scytta.key --domain local.scytta.com,localhost",
     "mkcert": "npm run mkcert:ca && npm run mkcert:test",
     "build_service": "docker build --target test -t gzp79/shine-identity:test --progress=plain ..",
+    "format": "prettier 'src/**/*.{ts,tsx,js}' --write",
     "jest": "jest -i --verbose=false",
-    "check:format": "prettier 'src/**/*.{ts,tsx,js}' --check",
-    "check": "npm run check:format",
-    "format": "prettier 'src/**/*.{ts,tsx,js}' --write"
+    "report": "xunit-viewer -r ./reports/junit.xml -o ./reports/index.html",
+    "test:regression": "npm run jest regression || npm run report",
+    "test:format": "prettier 'src/**/*.{ts,tsx,js}' --check",
+    "test": "npm run test:format && npm run test:regression"
   },
   "author": "",
   "license": "MIT",
@@ -20,6 +22,7 @@
     "node": ">=18.18.2"
   },
   "devDependencies": {
+    "@trivago/prettier-plugin-sort-imports": "^4.3.0",
     "@types/debug": "^4.1.10",
     "@types/express": "^4.17.17",
     "@types/jest": "^29.5.5",
@@ -29,7 +32,6 @@
     "@types/tough-cookie": "^4.0.3",
     "@types/uuid-validate": "^0.0.1",
     "cross-env": "^7.0.3",
-    "jest-stare": "^2.5.1",
     "mkcert": "^3.1.0",
     "prettier": "^3.0.3",
     "ts-jest": "^29.1.1",
@@ -44,11 +46,12 @@
     "jest": "^29.7.0",
     "jest-expect-message": "^1.1.3",
     "jest-extended": "^4.0.2",
+    "jest-junit": "^16.0.0",
     "joi": "^17.11.0",
     "superagent": "^8.1.2",
-    "superagent-logger": "^1.1.0",
     "tough-cookie": "^4.1.3",
     "ts-jose": "^4.15.4",
-    "uuid-validate": "^0.0.3"
+    "uuid-validate": "^0.0.3",
+    "xunit-viewer": "^10.5.1"
   }
 }
