+++ integration-test/regression/login_token.ts
@@ -1,6 +1,6 @@
-import request from 'superagent';
+import request from '$lib/request';
 import { getPageRedirectUrl } from '$lib/page_utils';
-import { UserInfo, getCookies, getUserInfo } from '$lib/auth_utils';
+import { UserInfo, getCookies, getUserInfo } from '$lib/user_utils';
 import config from '../test.config';
 import { Cookie } from 'tough-cookie';
 import { TestUser } from '$lib/user';
@@ -10,9 +10,7 @@ describe('Login with token', () => {
         const response = await request
             .get(config.getUrlFor('identity/auth/token/login'))
             .query({ rememberMe: 'invalid value' })
-            .send()
-            .catch((err) => err.response);
-
+            .send();
         expect(response.statusCode).toEqual(200);
         expect(getPageRedirectUrl(response.text)).toEqual(
             'https://web.sandbox.com:8080/error?type=invalidInput&status=400'
@@ -29,9 +27,7 @@ describe('Login with token', () => {
         const response = await request
             .get(config.getUrlFor('identity/auth/token/login'))
             .query(config.defaultRedirects)
-            .send()
-            .catch((err) => err.response);
-
+            .send();
         expect(response.statusCode).toEqual(200);
         expect(getPageRedirectUrl(response.text)).toEqual(config.defaultRedirects.loginUrl);
 
@@ -45,9 +41,7 @@ describe('Login with token', () => {
         const response = await request
             .get(config.getUrlFor('identity/auth/token/login'))
             .query({ rememberMe: false, ...config.defaultRedirects })
-            .send()
-            .catch((err) => err.response);
-
+            .send();
         expect(response.statusCode).toEqual(200);
         expect(getPageRedirectUrl(response.text)).toEqual(config.defaultRedirects.loginUrl);
 
@@ -61,9 +55,7 @@ describe('Login with token', () => {
         const response = await request
             .get(config.getUrlFor('identity/auth/token/login'))
             .query({ rememberMe: true, ...config.defaultRedirects })
-            .send()
-            .catch((err) => err.response);
-
+            .send();
         expect(response.statusCode).toEqual(200);
         expect(getPageRedirectUrl(response.text)).toEqual(config.defaultRedirects.redirectUrl);
 
@@ -84,9 +76,7 @@ describe('Login with token for returning user', () => {
         const response = await request
             .get(config.getUrlFor('identity/auth/token/login'))
             .query({ rememberMe: true, ...config.defaultRedirects })
-            .send()
-            .catch((err) => err.response);
-
+            .send();
         expect(response.statusCode).toEqual(200);
         expect(getPageRedirectUrl(response.text)).toEqual(config.defaultRedirects.redirectUrl);
 
@@ -106,9 +96,7 @@ describe('Login with token for returning user', () => {
             .get(config.getUrlFor('identity/auth/token/login'))
             .query({ rememberMe: true, ...config.defaultRedirects })
             .set('Cookie', [`sid=${cookies.sid.value}`])
-            .send()
-            .catch((err) => err.response);
-
+            .send();
         expect(response.statusCode).toEqual(200);
         expect(getPageRedirectUrl(response.text)).toEqual(
             config.defaultRedirects.errorUrl + '?type=logoutRequired&status=400'
@@ -128,9 +116,7 @@ describe('Login with token for returning user', () => {
             .get(config.getUrlFor('identity/auth/token/login'))
             .query({ rememberMe: true, ...config.defaultRedirects })
             .set('Cookie', [`sid=${cookies.sid.value}`, `tid=${cookies.tid.value}`])
-            .send()
-            .catch((err) => err.response);
-
+            .send();
         expect(response.statusCode).toEqual(200);
         expect(getPageRedirectUrl(response.text)).toEqual(
             config.defaultRedirects.errorUrl + '?type=logoutRequired&status=400'
@@ -151,9 +137,7 @@ describe('Login with token for returning user', () => {
             .get(config.getUrlFor('identity/auth/token/login'))
             .query({ ...config.defaultRedirects })
             .set('Cookie', [`tid=${cookies.tid.value}`])
-            .send()
-            .catch((err) => err.response);
-
+            .send();
         expect(response.statusCode).toEqual(200);
         expect(getPageRedirectUrl(response.text)).toEqual(config.defaultRedirects.redirectUrl);
 
@@ -171,9 +155,7 @@ describe('Login with token for returning user', () => {
             .get(config.getUrlFor('identity/auth/token/login'))
             .query({ rememberMe: false, ...config.defaultRedirects })
             .set('Cookie', [`tid=${cookies.tid.value}`])
-            .send()
-            .catch((err) => err.response);
-
+            .send();
         expect(response.statusCode).toEqual(200);
         expect(getPageRedirectUrl(response.text)).toEqual(config.defaultRedirects.redirectUrl);
 
@@ -191,9 +173,7 @@ describe('Login with token for returning user', () => {
             .get(config.getUrlFor('identity/auth/token/login'))
             .query({ rememberMe: true, ...config.defaultRedirects })
             .set('Cookie', [`tid=${cookies.tid.value}`])
-            .send()
-            .catch((err) => err.response);
-
+            .send();
         expect(response.statusCode).toEqual(200);
         expect(getPageRedirectUrl(response.text)).toEqual(config.defaultRedirects.redirectUrl);
 
@@ -232,8 +212,7 @@ describe('Login with token for returning user', () => {
                 .get(config.getUrlFor('identity/api/auth/user/info'))
                 .set('Cookie', [`sid=${user.sid}`])
                 .set({ ...site_info, ...mod })
-                .send()
-                .catch((err) => err.response);
+                .send();
             expect(response.statusCode).toEqual(401);
         }
     });
