+++ integration-test/src/expect/auth_exts.ts
@@ -1,0 +1,83 @@
+import { Cookie } from 'tough-cookie';
+import uuidValidate from 'uuid-validate';
+import { UserInfo } from '../user';
+import { intoMatcherResult } from './utils';
+
+interface CustomMatchers<R = unknown> {
+    toBeClearCookie(): R;
+    toBeValidTID(): R;
+    toBeValidSID(): R;
+    toBeValidEID(): R;
+    //toBeGuestUser(): R;
+}
+
+declare global {
+    namespace jest {
+        interface Expect extends CustomMatchers {}
+        interface Matchers<R> extends CustomMatchers<R> {}
+        interface InverseAsymmetricMatchers extends CustomMatchers {}
+    }
+}
+
+const matchers: jest.ExpectExtendMap = {
+    toBeClearCookie(received: Cookie) {
+        const expected = expect.objectContaining({
+            secure: true,
+            httpOnly: true,
+            sameSite: 'lax',
+            expires: expect.toBeBefore(new Date())
+        });
+        return intoMatcherResult(this, received, expected);
+    },
+
+    toBeValidTID(received: Cookie) {
+        const expected = expect.objectContaining({
+            key: 'tid',
+            secure: true,
+            httpOnly: true,
+            sameSite: 'lax',
+            path: '/identity/auth',
+            domain: 'cloud.sandbox.com',
+            expires: expect.toBeAfter(new Date())
+        });
+        return intoMatcherResult(this, received, expected);
+    },
+
+    toBeValidSID(received: Cookie) {
+        const expected = expect.objectContaining({
+            key: 'sid',
+            secure: true,
+            httpOnly: true,
+            sameSite: 'lax',
+            path: '/',
+            domain: 'sandbox.com',
+            expires: 'Infinity' //session scoped
+        });
+        return intoMatcherResult(this, received, expected);
+    },
+
+    toBeValidEID(received: Cookie) {
+        const expected = expect.objectContaining({
+            key: 'eid',
+            secure: true,
+            httpOnly: true,
+            sameSite: 'lax',
+            path: '/identity/auth',
+            domain: 'cloud.sandbox.com',
+            expires: 'Infinity'
+        });
+        return intoMatcherResult(this, received, expected);
+    }
+
+    /*toBeGuestUser(received: UserInfo) {
+        const expected = expect.objectContaining({
+            userId: expect.toSatisfy((id: any) => uuidValidate(id)),
+            name: expect.toStartWith('Freshman_'),
+            sessionLength: expect.not.toBeNegative(),
+            roles: []
+        });
+        return intoMatcherResult(this, received, expected);
+    }*/
+};
+
+export default matchers;
