+++ integration-test/src/utils.ts
@@ -18,10 +18,10 @@ export function joinURL(baseUrl: URL, path: string): string {
 export function parseSignedCookie(value: string): any {
     const json = decodeURIComponent(value);
     return JSON.parse(json.substring(44));
-};
+}
 
 export function getSHA256Hash(text: string): string {
     const hash = crypto.createHash('sha256');
     hash.update(text);
     return hash.digest('hex');
-  }
+}
