+++ service/server_config.json
@@ -28,6 +28,7 @@
         "ttlSession": 43200,
         "ttlAccessToken": 1209600,
         "ttlSingleAccess": 600,
+        "ttlApiKey": 31536000,
         "openid": {
             "google": {
                 "discoveryUrl": "https://accounts.google.com",
