+++ service/src/auth/token/token_generator.rs
@@ -1,4 +1,3 @@
-use chrono::Duration;
 use ring::rand::{SecureRandom, SystemRandom};
 use thiserror::Error as ThisError;
 
@@ -7,27 +6,13 @@ use thiserror::Error as ThisError;
 pub struct TokenGeneratorError(String);
 
 #[derive(Clone)]
-pub struct TokenGenerator {
-    ttl_access_token: Duration,
-    ttl_single_access: Duration,
-    random: SystemRandom,
+pub struct TokenGenerator<'a> {
+    random: &'a SystemRandom,
 }
 
-impl TokenGenerator {
-    pub fn new(ttl_access_token: Duration, ttl_single_access: Duration) -> Self {
-        Self {
-            ttl_access_token,
-            ttl_single_access,
-            random: SystemRandom::new(),
-        }
-    }
-
-    pub fn ttl_access_token(&self) -> Duration {
-        self.ttl_access_token
-    }
-
-    pub fn ttl_single_access(&self) -> Duration {
-        self.ttl_single_access
+impl<'a> TokenGenerator<'a> {
+    pub fn new(random: &'a SystemRandom) -> Self {
+        Self { random }
     }
 
     pub fn generate_token(&self) -> Result<String, TokenGeneratorError> {
